feat: create and update classes service, repository and route.
Implement get session method and improved auth security

+++ apps/api/app/api/classes/[id]/route.ts
@@ -1,11 +1,17 @@
 import { NextRequest, NextResponse } from 'next/server'
 import { classesService } from '@/services/classes.service'
+import { getSession } from '@/config/get-session'
 
 export async function GET(
   req: NextRequest,
   { params }: { params: Promise<{ id: string }> }
 ) {
   try {
+    const session = await getSession()
+    if (!session) {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+    }
+
     const { id } = await params
     const cls = await classesService.getById(Number(id))
     return NextResponse.json(cls)
@@ -19,6 +25,11 @@ export async function PUT(
   { params }: { params: Promise<{ id: string }> }
 ) {
   try {
+    const session = await getSession()
+    if (!session) {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+    }
+
     const { id } = await params
     const body = await req.json()
     const updatedClass = await classesService.update(Number(id), body)
@@ -33,6 +44,11 @@ export async function DELETE(
   { params }: { params: Promise<{ id: string }> }
 ) {
   try {
+    const session = await getSession()
+    if (!session) {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+    }
+
     const { id } = await params
     await classesService.delete(Number(id))
     return new NextResponse(null, { status: 204 })
